Update Course Enhancement/Course_enhancement_report_2019s1_program_info.py
The database was edited to account for Vertical studios

Code updated to reflect changes in database. I still need to look at instances where there are mulitple
VE clusters in a semester for the same course.
Best current idea is to check for match course_code_ces then matching course_codes and coalesce the results.